refactors test structure.
